requirements.txt
streamlit==1.41.1
requests==2.32.3
google-generativeai==0.8.3